Compose: Move K1 specific code to a separate package
Ideally, there should be no references from the Compose IR plugin to
K1 specific code. Currently, there are still two references from
`ComposableSymbolRemapper` and `CopyDefaultValuesFromExpectLowering`,
but these passes are K1 specific anyway and shouldn't run on K2.

The remaining `BindingContext` usages in `ComposeWritableSlices` don't
refer to any frontend data structures and are only used internally in
the Compose IR plugin. Technically, this is still a K1 specific API and
we should remove it, but it's harmless for now.
