+++ compose/compiler/compiler-hosted/src/main/java/androidx/compose/compiler/plugins/kotlin/ComposeFqNames.kt
@@ -16,13 +16,6 @@
 
 package androidx.compose.compiler.plugins.kotlin
 
-import org.jetbrains.kotlin.descriptors.ModuleDescriptor
-import org.jetbrains.kotlin.descriptors.SourceElement
-import org.jetbrains.kotlin.descriptors.annotations.Annotated
-import org.jetbrains.kotlin.descriptors.annotations.AnnotationDescriptor
-import org.jetbrains.kotlin.descriptors.annotations.Annotations
-import org.jetbrains.kotlin.descriptors.findClassAcrossModuleDependencies
-import org.jetbrains.kotlin.descriptors.impl.AnonymousFunctionDescriptor
 import org.jetbrains.kotlin.ir.declarations.IrAnnotationContainer
 import org.jetbrains.kotlin.ir.types.IrType
 import org.jetbrains.kotlin.ir.util.hasAnnotation
@@ -30,12 +23,6 @@ import org.jetbrains.kotlin.name.CallableId
 import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.name.Name
-import org.jetbrains.kotlin.resolve.constants.ConstantValue
-import org.jetbrains.kotlin.resolve.descriptorUtil.annotationClass
-import org.jetbrains.kotlin.types.KotlinType
-import org.jetbrains.kotlin.types.TypeUtils.NO_EXPECTED_TYPE
-import org.jetbrains.kotlin.types.TypeUtils.UNIT_EXPECTED_TYPE
-import org.jetbrains.kotlin.types.typeUtil.replaceAnnotations
 
 private const val root = "androidx.compose.runtime"
 private const val internalRoot = "$root.internal"
@@ -126,94 +113,11 @@ object ComposeFqNames {
     val Stable = fqNameFor("Stable")
     val Immutable = fqNameFor("Immutable")
     val Composer = ComposeClassIds.Composer.asSingleFqName()
-    val ComposeVersion = fqNameFor("ComposeVersion")
-    val Package = FqName(root)
     val StabilityInferred = ComposeClassIds.StabilityInferred.asSingleFqName()
-
-    fun makeComposableAnnotation(module: ModuleDescriptor): AnnotationDescriptor =
-        object : AnnotationDescriptor {
-            override val type: KotlinType
-                get() = module.findClassAcrossModuleDependencies(
-                    ClassId.topLevel(Composable)
-                )!!.defaultType
-            override val allValueArguments: Map<Name, ConstantValue<*>> get() = emptyMap()
-            override val source: SourceElement get() = SourceElement.NO_SOURCE
-            override fun toString() = "[@Composable]"
-        }
 }
-
-private fun makeComposableAnnotation(module: ModuleDescriptor): AnnotationDescriptor =
-    object : AnnotationDescriptor {
-        override val type: KotlinType
-            get() = module.findClassAcrossModuleDependencies(
-                ComposeClassIds.Composable
-            )!!.defaultType
-        override val allValueArguments: Map<Name, ConstantValue<*>> get() = emptyMap()
-        override val source: SourceElement get() = SourceElement.NO_SOURCE
-        override fun toString() = "[@Composable]"
-    }
-
-fun KotlinType.makeComposable(module: ModuleDescriptor): KotlinType {
-    if (hasComposableAnnotation()) return this
-    val annotation = makeComposableAnnotation(module)
-    return replaceAnnotations(Annotations.create(annotations + annotation))
-}
-
-fun AnonymousFunctionDescriptor.annotateAsComposable(module: ModuleDescriptor) =
-    AnonymousFunctionDescriptor(
-        containingDeclaration,
-        Annotations.create(annotations + makeComposableAnnotation(module)),
-        kind,
-        source,
-        isSuspend
-    )
 
 fun IrType.hasComposableAnnotation(): Boolean =
     hasAnnotation(ComposeFqNames.Composable)
 
 fun IrAnnotationContainer.hasComposableAnnotation(): Boolean =
     hasAnnotation(ComposeFqNames.Composable)
-
-fun KotlinType.hasComposableAnnotation(): Boolean =
-    !isSpecialType && annotations.findAnnotation(ComposeFqNames.Composable) != null
-fun Annotated.hasComposableAnnotation(): Boolean =
-    annotations.findAnnotation(ComposeFqNames.Composable) != null
-fun Annotated.hasNonRestartableComposableAnnotation(): Boolean =
-    annotations.findAnnotation(ComposeFqNames.NonRestartableComposable) != null
-fun Annotated.hasReadonlyComposableAnnotation(): Boolean =
-    annotations.findAnnotation(ComposeFqNames.ReadOnlyComposable) != null
-fun Annotated.hasExplicitGroupsAnnotation(): Boolean =
-    annotations.findAnnotation(ComposeFqNames.ExplicitGroupsComposable) != null
-fun Annotated.hasDisallowComposableCallsAnnotation(): Boolean =
-    annotations.findAnnotation(ComposeFqNames.DisallowComposableCalls) != null
-fun Annotated.compositionTarget(): String? =
-    annotations.map { it.compositionTarget() }.firstOrNull { it != null }
-
-fun Annotated.hasCompositionTargetMarker(): Boolean =
-    annotations.findAnnotation(
-        ComposeFqNames.ComposableTargetMarker
-    ) != null
-
-fun AnnotationDescriptor.compositionTarget(): String? =
-    if (fqName == ComposeFqNames.ComposableTarget)
-        allValueArguments[ComposeFqNames.ComposableTargetApplierArgument]?.value as? String
-    else if (annotationClass?.hasCompositionTargetMarker() == true) this.fqName.toString() else null
-
-fun Annotated.compositionScheme(): String? =
-    annotations.findAnnotation(
-        ComposeFqNames.ComposableInferredTarget
-    )?.allValueArguments?.let {
-        it[ComposeFqNames.ComposableInferredTargetSchemeArgument]?.value as? String
-    }
-
-fun Annotated.compositionOpenTarget(): Int? =
-    annotations.findAnnotation(
-        ComposeFqNames.ComposableOpenTarget
-    )?.allValueArguments?.let {
-        it[ComposeFqNames.ComposableOpenTargetIndexArgument]?.value as Int
-    }
-
-internal val KotlinType.isSpecialType: Boolean get() =
-    this === NO_EXPECTED_TYPE || this === UNIT_EXPECTED_TYPE
-
-val AnnotationDescriptor.isComposableAnnotation: Boolean get() = fqName == ComposeFqNames.Composable
