+++ compose/compiler/compiler-hosted/src/main/java/androidx/compose/compiler/plugins/kotlin/k1/ComposeTypeResolutionInterceptorExtension.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 The Android Open Source Project
+ * Copyright 2023 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,8 @@
  * limitations under the License.
  */
 
-package androidx.compose.compiler.plugins.kotlin
+package androidx.compose.compiler.plugins.kotlin.k1
 
-import androidx.compose.compiler.plugins.kotlin.analysis.ComposeWritableSlices
-import androidx.compose.compiler.plugins.kotlin.analysis.ComposeWritableSlices.INFERRED_COMPOSABLE_DESCRIPTOR
 import org.jetbrains.kotlin.descriptors.impl.AnonymousFunctionDescriptor
 import org.jetbrains.kotlin.extensions.internal.TypeResolutionInterceptorExtension
 import org.jetbrains.kotlin.psi.KtElement
@@ -48,7 +46,7 @@ open class ComposeTypeResolutionInterceptorExtension : TypeResolutionInterceptor
             // If the expected type has an @Composable annotation then the literal function
             // expression should infer a an @Composable annotation
             descriptor.annotateAsComposable(context.scope.ownerDescriptor.module).also {
-                context.trace.record(INFERRED_COMPOSABLE_DESCRIPTOR, it, true)
+                context.trace.record(FrontendWritableSlices.INFERRED_COMPOSABLE_DESCRIPTOR, it, true)
             }
         } else {
             descriptor
@@ -67,7 +65,7 @@ open class ComposeTypeResolutionInterceptorExtension : TypeResolutionInterceptor
             element.getAnnotationEntries().hasComposableAnnotation(context.trace.bindingContext) ||
             context.hasComposableExpectedType(element)
         ) {
-            context.trace.record(ComposeWritableSlices.INFERRED_COMPOSABLE_LITERAL, element, true)
+            context.trace.record(FrontendWritableSlices.INFERRED_COMPOSABLE_LITERAL, element, true)
             return resultType.makeComposable(context.scope.ownerDescriptor.module)
         }
         return resultType
