+++ compose/compiler/compiler-hosted/src/main/resources/META-INF/services/org.jetbrains.kotlin.diagnostics.rendering.DefaultErrorMessages$Extension
@@ -1,1 +1,1 @@
-androidx.compose.compiler.plugins.kotlin.ComposeErrorMessages
+androidx.compose.compiler.plugins.kotlin.k1.ComposeErrorMessages
